# docs/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/sources": {
      "get": {
        "operationId": "SourcesController_list",
        "parameters": [
          {
            "name": "sourceType",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "sources"
        ]
      },
      "post": {
        "operationId": "SourcesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSourceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "sources"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/sources/{id}": {
      "get": {
        "operationId": "SourcesController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "sources"
        ]
      }
    },
    "/v1/sources/{id}/comments": {
      "get": {
        "operationId": "SourcesController_getComments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from a previous page.nextCursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "sources"
        ]
      }
    },
    "/v1/sources/{id}/documents/{documentId}": {
      "get": {
        "operationId": "SourcesController_getDocument",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "sources"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/sources/{id}/documents": {
      "post": {
        "operationId": "SourcesController_addDocument",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSourceDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "sources"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/sources/{id}/archive": {
      "patch": {
        "operationId": "SourcesController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArchiveSourceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "sources"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Dau Viet API",
    "description": "Living Digital Atlas of Vietnam - backend API contract for web, mobile, and admin clients. See docs/backend/BACKEND_HANDOFF.md for the full handoff contract this document is generated alongside.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateSourceDto": {
        "type": "object",
        "properties": {
          "sourceType": {
            "type": "string",
            "enum": [
              "BOOK",
              "PRIMARY_DOCUMENT",
              "ARCHIVAL_DOCUMENT",
              "MAP",
              "MANUSCRIPT",
              "ACADEMIC_PAPER",
              "MUSEUM_RECORD",
              "GOVERNMENT_DOCUMENT",
              "UNESCO_RECORD",
              "NEWSPAPER",
              "PHOTO_ARCHIVE",
              "ORAL_HISTORY",
              "WEBSITE",
              "OTHER"
            ]
          },
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "organization": {
            "type": "string"
          },
          "publisher": {
            "type": "string"
          },
          "publicationYear": {
            "type": "number"
          },
          "isbn": {
            "type": "string"
          },
          "issn": {
            "type": "string"
          },
          "edition": {
            "type": "string"
          },
          "volume": {
            "type": "string"
          },
          "archiveName": {
            "type": "string"
          },
          "archiveCode": {
            "type": "string"
          },
          "originalLanguage": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "accessedAt": {
            "type": "string"
          },
          "credibilityLevel": {
            "type": "string",
            "enum": [
              "PRIMARY",
              "SECONDARY",
              "TERTIARY",
              "UNKNOWN"
            ]
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "sourceType",
          "title"
        ]
      },
      "CreateSourceDocumentDto": {
        "type": "object",
        "properties": {
          "mediaAssetId": {
            "type": "string"
          },
          "pageCount": {
            "type": "number"
          },
          "extractedText": {
            "type": "string"
          },
          "usageRights": {
            "type": "string"
          },
          "rightsHolder": {
            "type": "string"
          },
          "accessPolicy": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PREVIEW_ONLY",
              "METADATA_ONLY",
              "RESTRICTED"
            ]
          }
        },
        "required": [
          "mediaAssetId"
        ]
      },
      "ArchiveSourceDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}